requirements.txt
# Time Graph Uygulaması - Gerekli Paketler
# ==========================================

# GUI Framework
PyQt5>=5.15.0

# Veri İşleme
pandas>=1.3.0
numpy>=1.21.0
vaex>=4.0.0

# Görselleştirme
matplotlib>=3.5.0

# EXE Build (geliştirme için)
pyinstaller>=5.0.0
pillow>=8.0.0

# Ek Kütüphaneler
openpyxl>=3.0.0  # Excel dosya desteği için
